% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw_mean_col.R
\name{bw_mean_col}
\alias{bw_mean_col}
\title{Title}
\usage{
bw_mean_col(df, x, by, verbose = FALSE, saveplot = "", ...)
}
\arguments{
\item{df}{A dataframe.}

\item{x}{The data column within \code{df} for which means are to be calculated.}

\item{by}{The data column within \code{df} containing the variable by which the columns are grouped.}

\item{verbose}{Optional argument that, when \code{TRUE}, prints information on the function's progress.}

\item{saveplot}{Optional argument to be used to save the output to file. It is the name the file should be saved as.}

\item{...}{Optional arguments for the mean calculation - see \code{mean()} for more information.}
}
\value{
A \code{ggplot} object containing a bar graph of mean \code{x} per \code{by} with specialized theme settings.
}
\description{
Title
}
\examples{
bw_mean_col(iris, Sepal.Width, Species)
bw_mean_col(mtcars, mpg, cyl, trim = 0.2)

}
